Make certain lines thinner.
SVG lines look like garbage once the stroke width is less than 1 because of aliasing. Change how the shape is rendered to achieve a crisp but not faded line look. Note that this works well for my screen, but could look pretty crummy on stuff that's less high-res. If there's time, replace the elements that need finer rendering with divs.