Fix issue with start being drawn before a series
There's a weird issue where the star is getting appended to the parent element BEFORE some of the circles/lines. This happened a few times with the first chart in early builds, but once I started building out other elements after each series (like the key), it disappeared. This indicates some sort of race-condition, so to get around this issue, if there isn't anything drawn after the completed series, I'm just going to add a fake element, to avoid the problem.