Adding ability to reinitialize the chart.
This is a relatively expensive operation for the first chart - with the current set up, this requires us to update 90% of the child nodes. Future improvements definitely include figuring out if there's a more efficient way to do this. In the mean time, stamp necessary initialization data onto the node so we reduce the amount of recalculating we need to do to reset the viz.

Also removing commented out code. We don't need it!